# Improve performance of NewtonsoftJsonServiceSerializer.Clone